api/requirements.txt
fastapi
uvicorn
python-multipart
web3
requests
facebook-business
patreon
beautifulsoup4
httpx
stripe
python-dotenv
google-cloud-storage
